requirements.txt
streamlit
tensorflow
pillow
werkzeug
scikit-learn
pickle
pandas
numpy
scikit-learn-intelex